feat(db): configure H2 + Flyway for patients table, seed sample records, and enforce input validation
Main issue: previously there was no data input validation, allowing creation of records with null values.
This commit adds NOT NULL constraints and Bean Validation annotations to prevent invalid patient data.

configure H2 + JPA ddl-auto, add Flyway migration for patients table, and seed with 3 sample records
[
  {
    "firstName": "Alice",
    "middleName": "Marie",
    "lastName": "Johnson",
    "age": 32,
    "email": "alice.johnson@example.com"
  },
  {
    "firstName": "Bruno",
    "middleName": "Miguel",
    "lastName": "Silva",
    "age": 45,
    "email": "bruno.silva@example.com"
  },
  {
    "firstName": "Carla",
    "middleName": "Fernanda",
    "lastName": "Santos",
    "age": 28,
    "email": "carla.santos@example.com"
  }
]

CREATE TABLE patients (
    id BIGINT AUTO_INCREMENT PRIMARY KEY,
    first_name VARCHAR(255),
    middle_name VARCHAR(255),
    last_name VARCHAR(255),
    age INT,
    email VARCHAR(255)
);
