+++ .idea/compiler.xml
@@ -2,10 +2,14 @@
 <project version="4">
   <component name="CompilerConfiguration">
     <annotationProcessing>
+      <profile default="true" name="Default" enabled="true" />
       <profile name="Maven default annotation processors profile" enabled="true">
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
+        <module name="medical-service-one" />
+        <module name="shop-service" />
+        <module name="coffee-service" />
         <module name="spring-transaction" />
       </profile>
     </annotationProcessing>
@@ -13,7 +17,11 @@
   </component>
   <component name="JavacSettings">
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
-      <module name="coffeeservice" options="-parameters -parameters" />
+      <module name="coffee-service" options="-parameters -parameters" />
+      <module name="marketing" options="-parameters -parameters" />
+      <module name="medical-service" options="-parameters -parameters" />
+      <module name="medical-service-one" options="-parameters -parameters" />
+      <module name="shop-service" options="-parameters -parameters" />
       <module name="spring-transaction" options="-parameters -parameters" />
     </option>
   </component>
