+++ .idea/dictionaries/project.xml
@@ -2,6 +2,7 @@
   <dictionary name="project">
     <words>
       <w>coffeeservice</w>
+      <w>edosnwade</w>
       <w>muhongo</w>
       <w>springtx</w>
       <w>vanilson</w>
