+++ .idea/encodings.xml
@@ -1,6 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/coffee-service/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/medical-service/medical-service-one/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/medical-service/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/medical-service/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/shop-service/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/spring-transaction/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
