+++ coffee-service/src/main/java/code/with/vanilson/CoffeeserviceApplication.java
@@ -1,0 +1,13 @@
+package code.with.vanilson;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CoffeeserviceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CoffeeserviceApplication.class, args);
+    }
+
+}
