+++ coffee-service/src/main/java/code/with/vanilson/repository/CoffeeRepository.java
@@ -1,0 +1,10 @@
+package code.with.vanilson.repository;
+
+import code.with.vanilson.model.Coffee;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface CoffeeRepository extends JpaRepository<Coffee, Long> {
+    List<Coffee> findByName(String name);
+}
