+++ coffee-service/src/main/resources/schema.sql
