+++ coffee-service/src/test/java/code/with/vanilson/CoffeeServiceTestConfig.java
@@ -1,0 +1,27 @@
+package code.with.vanilson;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+import org.springframework.context.annotation.Profile;
+import org.springframework.jdbc.datasource.DriverManagerDataSource;
+
+import javax.sql.DataSource;
+
+@Configuration
+@Profile("test")
+public class CoffeeServiceTestConfig {
+    @Primary
+    @Bean
+    public DataSource dataSource() {
+
+        //  Setup a data source for our tests
+        DriverManagerDataSource dataSource = new DriverManagerDataSource();
+        dataSource.setDriverClassName("org.h2.Driver");
+        dataSource.setUrl("jdbc:h2:mem:db;DB_CLOSE_DELAY=-1");
+        dataSource.setUsername("sa");
+        dataSource.setPassword("");
+
+        return dataSource;
+    }
+}
