+++ coffee-service/src/test/java/code/with/vanilson/CoffeeserviceApplicationTests.java
@@ -1,0 +1,12 @@
+package code.with.vanilson;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CoffeeserviceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+}
