+++ medical-service/medical-service-one/src/main/java/code/with/vanilson/MedicalApplication.java
@@ -1,0 +1,18 @@
+/**
+ * Author: vanilson muhongo
+ * Date:2025-11-28
+ * Time:15:40
+ * Version:1
+ */
+
+package code.with.vanilson;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MedicalApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(MedicalApplication.class, args);
+    }
+}
