+++ medical-service/medical-service-one/src/main/java/code/with/vanilson/patients/PatientRequest.java
@@ -1,0 +1,29 @@
+package code.with.vanilson.patients;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Data
+public class PatientRequest {
+
+    private String firstName;
+    private String middleName;
+    private String lastName;
+    private Integer age;
+    private String email;
+
+    public Patient toEntity() {
+        return Patient.builder()
+                .firstName(firstName)
+                .middleName(middleName)
+                .lastName(lastName)
+                .age(age)
+                .email(email)
+                .build();
+    }
+}
