+++ medical-service/medical-service-one/src/main/java/code/with/vanilson/patients/PatientsRepository.java
@@ -1,0 +1,12 @@
+package code.with.vanilson.patients;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface PatientsRepository extends CrudRepository<Patient, Long> {
+
+    List<Patient> findAll();
+}
