+++ medical-service/medical-service-one/src/main/resources/application.yaml
@@ -1,0 +1,17 @@
+spring:
+  datasource:
+    url: jdbc:h2:mem:medical-facility
+    username: sa
+    password: password
+    driverClassName: org.h2.Driver
+  jpa:
+    spring.jpa.database-platform: org.hibernate.dialect.H2Dialect
+    hibernate:
+      ddl-auto: update   # ou create-drop
+      show-sql: true
+
+flyway:
+  user: sa
+  password: password
+  schemas: med-facility
+  url: jdbc:h2:mem:medical-facility
