+++ medical-service/pom.xml
@@ -1,0 +1,18 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>code.with.vanilson</groupId>
+        <artifactId>marketing</artifactId>
+        <version>0.0.1-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>medical-service</artifactId>
+    <!-- not to create pom-->
+    <packaging>pom</packaging>
+    <modules>
+        <module>medical-service-one</module>
+    </modules>
+
+
+</project>
