+++ pom.xml
@@ -9,12 +9,14 @@
         <relativePath/> <!-- lookup parent from a repository -->
     </parent>
     <groupId>code.with.vanilson</groupId>
-    <artifactId>coffeeservice</artifactId>
+    <artifactId>marketing</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <packaging>pom</packaging>
-    <name>coffeeservice</name>
     <modules>
         <module>spring-transaction</module>
+        <module>shop-service</module>
+        <module>coffee-service</module>
+        <module>medical-service</module>
     </modules>
     <properties>
         <java.version>17</java.version>
@@ -37,6 +39,17 @@
             <groupId>com.h2database</groupId>
             <artifactId>h2</artifactId>
             <scope>runtime</scope>
+        </dependency>
+        <!-- Actuator for health, info, metrics endpoints -->
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-actuator</artifactId>
+        </dependency>
+
+        <!-- Micrometer Prometheus registry -->
+        <dependency>
+            <groupId>io.micrometer</groupId>
+            <artifactId>micrometer-registry-prometheus</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -95,8 +108,23 @@
     <dependencyManagement>
         <dependencies>
             <dependency>
-                <groupId>code.with.vanilson</groupId>
+                <groupId>${project.groupId}</groupId>
                 <artifactId>spring-transaction</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>${project.groupId}</groupId>
+                <artifactId>coffee-service</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>${project.groupId}</groupId>
+                <artifactId>shop-service</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>${project.groupId}</groupId>
+                <artifactId>medical-service</artifactId>
                 <version>${project.version}</version>
             </dependency>
         </dependencies>
