+++ shop-service/src/main/java/code/with/vanilson/App.java
@@ -1,0 +1,11 @@
+package code.with.vanilson;
+
+/**
+ * Hello world!
+ *
+ */
+public class App {
+    public static void main(String[] args) {
+        System.out.println("Hello World!");
+    }
+}
