--- src/main/java/code/with/vanilson/coffeeservice/CoffeeServiceConfig.java
@@ -1,33 +1,0 @@
-package code.with.vanilson.coffeeservice;
-
-import javax.sql.DataSource;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.context.annotation.Profile;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-
-@Configuration
-@Profile("prod")
-public class CoffeeServiceConfig {
-    @Primary
-    @Bean
-    public DataSource dataSource() {
-
-        // Setup a data source for our tests
-        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setDriverClassName("org.h2.Driver");
-        dataSource.setUrl("jdbc:h2:mem:db;DB_CLOSE_DELAY=-1");
-        dataSource.setUsername("sa");
-        dataSource.setPassword("");
-
-//        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-//        dataSource.setDriverClassName("com.mysql.cj.jdbc.Driver");
-//        dataSource.setUrl("jdbc:mysql://localhost:3306/coffee");
-//        dataSource.setUsername("root");
-//        dataSource.setPassword("");
-
-        return dataSource;
-    }
-}
