--- src/main/java/code/with/vanilson/coffeeservice/CoffeeserviceApplication.java
@@ -1,13 +1,0 @@
-package code.with.vanilson.coffeeservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CoffeeserviceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CoffeeserviceApplication.class, args);
-	}
-
-}
