--- src/main/java/code/with/vanilson/coffeeservice/model/Coffee.java
@@ -1,79 +1,0 @@
-package code.with.vanilson.coffeeservice.model;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-import java.util.Objects;
-
-@Entity
-public class Coffee {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-    private String name;
-    private int version;
-
-    protected Coffee() {
-    }
-
-    public Coffee(String name) {
-        this.name = name;
-    }
-
-    public Coffee(String name, int version) {
-        this.name = name;
-        this.version = version;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getVersion() {
-        return version;
-    }
-
-    public void setVersion(int version) {
-        this.version = version;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (o == null || getClass() != o.getClass()) {return false;}
-
-        Coffee coffee = (Coffee) o;
-        return version == coffee.version && Objects.equals(id, coffee.id) &&
-                Objects.equals(name, coffee.name);
-    }
-
-    @Override
-    public int hashCode() {
-        int result = Objects.hashCode(id);
-        result = 31 * result + Objects.hashCode(name);
-        result = 31 * result + version;
-        return result;
-    }
-
-    @Override
-    public String toString() {
-        return "Coffee{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", version=" + version +
-                '}';
-    }
-}
