--- src/main/resources/application.properties
@@ -1,13 +1,0 @@
-# H2 Database Configuration
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=
-
-# Hibernate Configuration
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-spring.jpa.hibernate.ddl-auto=update
-
-# Enable H2 Console
-spring.h2.console.enabled=true
-spring.h2.console.path=/h2-console
