--- src/test/java/code/with/vanilson/coffeeservice/CoffeeserviceApplicationTests.java
@@ -1,12 +1,0 @@
-package code.with.vanilson.coffeeservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CoffeeserviceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-}
